requirements.txt
hydra-core
pandas
bertopic
sklearn